fix: cast explicito en el sembrado de tramos_precio
El insert mezclaba enteros con null en un VALUES, y Postgres no siempre puede
inferir el tipo de la columna: fallaba al sembrar los tramos por defecto.
Se agrega null::int.

Migraciones ya aplicadas en KSE. Se registran dos pendientes que surgieron al
aplicarlas: regenerar types/database.ts (las tablas de cartera no estan
tipadas y la pagina usa un cast temporal) y agregar politicas RLS, porque la
migracion otorgo permisos a 'authenticated' y hoy cualquier usuario con sesion
podria leer y escribir contratos y pagos.

+++ supabase/migraciones/20260912_cartera.sql
@@ -75,7 +75,7 @@ create unique index if not exists idx_tramo_abierto_unico
   on public.tramos_precio ((hasta is null)) where hasta is null;
 
 insert into public.tramos_precio (hasta, precio_usuario)
-select * from (values (1, 1200), (5, 1000), (15, 850), (30, 700), (null, 600)) as v(hasta, precio)
+select * from (values (1, 1200), (5, 1000), (15, 850), (30, 700), (null::int, 600)) as v(hasta, precio)
 where not exists (select 1 from public.tramos_precio);
 
 grant select, insert, update, delete on public.tramos_precio to service_role, authenticated;
